requirements.txt
web3>=6.18.0
aiohttp>=3.9.5
PyYAML>=6.0.1
pandas>=2.2.2
requests>=2.31.0
python-dotenv>=1.0.0